[qa] Clear workspace clippy warnings (clippy -D warnings green)
`cargo clippy --workspace --all-targets -- -D warnings` was red from
toolchain lint drift. Now green. Mostly cargo-clippy --fix auto-fixes
(needless borrows, is_multiple_of, same-type casts, clone-on-Copy,
useless vec!, redundant pattern matching, unnecessary parens). Hand fixes:
- loop-index → iterators (loop_arch_bench next_fire/cores);
- RefCell borrow no longer held across .await (loop_arch_bench submit ring);
- dropped redundant drop() / added `let _ =` on must-use Result;
- test file headers /// → //! (6 rsx-risk tests);
- std::slice::from_ref instead of &[x.clone()].
Two frozen rsx-cast/src/cast.rs changes, both behavior-identical: NAK
match→if, `>= y+1`→`> y`. Three justified #[allow]s with reasons: bench
doc_lazy_continuation (non-rendered bench docs, ×2), intentional early
drop of PG clients, and fixed-point `<int>_<4dp>` digit grouping.

+++ rsx-matching/tests/config_poll_test.rs
@@ -213,7 +213,7 @@ async fn write_applied_config_inserts_new() {
     let (_c, client) = setup_db().await;
     let symbol_id = 1u32;
     let now = now_ms();
-    let ts_ns = (now * 1_000_000) as u64;
+    let ts_ns = now * 1_000_000 ;
 
     let cfg = ScheduledConfig {
         config_version: 1,
@@ -244,7 +244,7 @@ async fn write_applied_config_updates_existing() {
     let (_c, client) = setup_db().await;
     let symbol_id = 1u32;
     let now = now_ms();
-    let ts_ns = (now * 1_000_000) as u64;
+    let ts_ns = now * 1_000_000 ;
 
     let cfg1 = ScheduledConfig {
         config_version: 1,
